+++ core/code/code.go
@@ -16,6 +16,7 @@ const (
 	TxFromSenderAlreadyInMempool uint32 = 113
 	TooLowGasPrice               uint32 = 114
 	WrongChainID                 uint32 = 115
+	CoinReserveUnderflow         uint32 = 116
 
 	// coin creation
 	CoinAlreadyExists uint32 = 201