+++ core/transaction/redeem_check.go
@@ -155,6 +155,12 @@ func (data RedeemCheckData) Run(tx *Transaction, context *state.State, isCheck b
 
 	if !decodedCheck.GasCoin.IsBaseCoin() {
 		coin := context.Coins.GetCoin(decodedCheck.GasCoin)
+		err := coin.CheckReserveUnderflow(commissionInBaseCoin)
+		if err != nil {
+			return Response{
+				Code: code.CoinReserveUnderflow,
+				Log:  err.Error()}
+		}
 		commission = formula.CalculateSaleAmount(coin.Volume(), coin.Reserve(), coin.Crr(), commissionInBaseCoin)
 	}