+++ core/transaction/sell_coin.go
@@ -302,6 +302,13 @@ func (data SellCoinData) Run(tx *Transaction, context *state.State, isCheck bool
 		}
 	}
 
+	err := checkConversionsReserveUnderflow(conversions, context)
+	if err != nil {
+		return Response{
+			Code: code.CoinReserveUnderflow,
+			Log:  err.Error()}
+	}
+
 	if !isCheck {
 		for _, ts := range totalSpends {
 			context.Accounts.SubBalance(sender, ts.Coin, ts.Value)