+++ core/transaction/send.go
@@ -28,6 +28,13 @@ func (data SendData) TotalSpend(tx *Transaction, context *state.State) (TotalSpe
 	if !tx.GasCoin.IsBaseCoin() {
 		coin := context.Coins.GetCoin(tx.GasCoin)
 
+		err := coin.CheckReserveUnderflow(commissionInBaseCoin)
+		if err != nil {
+			return nil, nil, nil, &Response{
+				Code: code.CoinReserveUnderflow,
+				Log:  err.Error()}
+		}
+
 		if coin.Reserve().Cmp(commissionInBaseCoin) < 0 {
 			return nil, nil, nil, &Response{
 				Code: code.CoinReserveNotSufficient,