+++ CHANGELOG.md
@@ -13,6 +13,8 @@ BREAKING CHANGES
 - [core] Fix sell coin commission calculation
 - [core] Enable multisignatures
 - [core] Do not delete coins with small reserve
+- [core] Do now allow to sell coins with reserve less than 10,000 bip
+- [core] Set min coin reserve to 10,000 bip
 - [gui] Remove GUI
 - [config] KeepStateHistory -> KeepLastStates
 - [config] Add state_cache_size option